Removal of mock project, added webviewlocator service
