+++ Sidekick.sln
@@ -23,8 +23,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{E4920C
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{1D8BF515-81E9-4769-9B27-24DFB8D8BDCC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sidekick.Mock", "src\Sidekick.Mock\Sidekick.Mock.csproj", "{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sidekick.Common.Blazor", "src\Sidekick.Common.Blazor\Sidekick.Common.Blazor.csproj", "{A78010BD-2B6B-4D13-A1A5-F90B1C395C1A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sidekick.Common", "src\Sidekick.Common\Sidekick.Common.csproj", "{484EAA27-27A0-4B0E-BA43-B5047D697CAE}"
@@ -79,14 +77,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}.Linux-Debug|Any CPU.ActiveCfg = Linux-Debug|Any CPU
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}.Linux-Debug|Any CPU.Build.0 = Linux-Debug|Any CPU
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}.Linux-Release|Any CPU.ActiveCfg = Linux-Release|Any CPU
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}.Linux-Release|Any CPU.Build.0 = Linux-Release|Any CPU
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A78010BD-2B6B-4D13-A1A5-F90B1C395C1A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A78010BD-2B6B-4D13-A1A5-F90B1C395C1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A78010BD-2B6B-4D13-A1A5-F90B1C395C1A}.Linux-Debug|Any CPU.ActiveCfg = Linux-Debug|Any CPU
@@ -260,7 +250,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{1C2974D8-B04E-4BF5-8CD8-65A82EE89493} = {1D8BF515-81E9-4769-9B27-24DFB8D8BDCC}
 		{A78010BD-2B6B-4D13-A1A5-F90B1C395C1A} = {E4920CFB-0EED-44B1-AC8D-8F5ECBCB7788}
 		{484EAA27-27A0-4B0E-BA43-B5047D697CAE} = {E4920CFB-0EED-44B1-AC8D-8F5ECBCB7788}
 		{3A9B57B9-08B5-4CA7-9B24-DAFB2E3438FF} = {E4920CFB-0EED-44B1-AC8D-8F5ECBCB7788}
