+++ src/Sidekick.Common.Platform/StartupExtensions.cs
@@ -1,6 +1,7 @@
 using System.Runtime.InteropServices;
 using Microsoft.Extensions.DependencyInjection;
 using Sidekick.Common.Platform.Clipboard;
+using Sidekick.Common.Platform.GameLogs;
 using Sidekick.Common.Platform.Interprocess;
 using Sidekick.Common.Platform.Keyboards;
 using Sidekick.Common.Platform.Localization;
@@ -33,6 +34,7 @@ namespace Sidekick.Common.Platform
 
             services.AddSidekickInitializableService<IKeyboardProvider, KeyboardProvider>();
             services.AddSingleton<IInterprocessService, InterprocessService>();
+            services.AddSingleton<IGameLogProvider, GameLogProvider>();
 
             return services;
         }
