+++ src/Sidekick.Common.Ui/App/AppMinimize.razor
@@ -1,3 +1,8 @@
+@if (!ViewLocator.SupportsMinimize)
+{
+    return;
+}
+
 <button type="button"
         @onclick="OnClick"
         @onclick:preventDefault="true"
