+++ src/Sidekick.Common/Sidekick.Common.csproj
@@ -9,6 +9,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.2" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.2" />
+    <PackageReference Include="Microsoft.Extensions.Localization" Version="8.0.11" />
     <PackageReference Include="Microsoft.Extensions.Localization.Abstractions" Version="8.0.11" />
     <PackageReference Include="Romanization.NET" Version="0.9.5" />
     <PackageReference Include="Serilog" Version="4.2.0" />
