--- src/Sidekick.Mock/MockApplicationService.cs
@@ -1,14 +1,0 @@
-using System;
-using System.Threading.Tasks;
-using Sidekick.Common.Platform;
-
-namespace Sidekick.Mock
-{
-    public class MockApplicationService : IApplicationService
-    {
-        public void Shutdown()
-        {
-            Environment.Exit(Environment.ExitCode);
-        }
-    }
-}
