--- src/Sidekick.Mock/MockTrayProvider.cs
@@ -1,12 +1,0 @@
-using System.Collections.Generic;
-using Sidekick.Common.Platform;
-
-namespace Sidekick.Mock
-{
-    public class MockTrayProvider : ITrayProvider
-    {
-        public void Initialize(List<TrayMenuItem> trayMenuItems)
-        {
-        }
-    }
-}
