+++ src/Sidekick.Modules.Chat/Keybinds/ChatKeybindHandler.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Logging;
-using Sidekick.Common.Game.GameLogs;
 using Sidekick.Common.Keybinds;
 using Sidekick.Common.Platform;
+using Sidekick.Common.Platform.GameLogs;
 using Sidekick.Common.Settings;
 
 namespace Sidekick.Modules.Chat.Keybinds
