+++ src/Sidekick.Wpf/Sidekick.Wpf.csproj
@@ -40,7 +40,6 @@
   <ItemGroup>
     <ProjectReference Include="..\Sidekick.Common.Updater\Sidekick.Common.Updater.csproj" />
     <ProjectReference Include="..\Sidekick.Common.Platform\Sidekick.Common.Platform.csproj" />
-    <ProjectReference Include="..\Sidekick.Mock\Sidekick.Mock.csproj" />
     <ProjectReference Include="..\Sidekick.Modules.Chat\Sidekick.Modules.Chat.csproj" />
     <ProjectReference Include="..\Sidekick.Modules.Development\Sidekick.Modules.Development.csproj" />
     <ProjectReference Include="..\Sidekick.Modules.General\Sidekick.Modules.General.csproj" />
