+++ tests/Sidekick.Apis.Poe.Tests/Sidekick.Apis.Poe.Tests.csproj
@@ -30,8 +30,8 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\Sidekick.Apis.PoeNinja\Sidekick.Apis.PoeNinja.csproj" />
     <ProjectReference Include="..\..\src\Sidekick.Apis.PoeWiki\Sidekick.Apis.PoeWiki.csproj" />
+    <ProjectReference Include="..\..\src\Sidekick.Apis.Poe\Sidekick.Apis.Poe.csproj" />
     <ProjectReference Include="..\..\src\Sidekick.Common\Sidekick.Common.csproj" />
-    <ProjectReference Include="..\..\src\Sidekick.Mock\Sidekick.Mock.csproj" />
   </ItemGroup>
 
 </Project>
